services/prank_shop/requirements.txt
python-multipart
uvicorn[standart]
pyjwt
fastapi